Create calculator functionality
created 2 variables to hold the numbers for the operation. Created event
listenrs on all number buttons, once they are clicked the value will be
stored in num1.

Created event listenrs for the operator buttons, once they are clicked,
the display will clear for the second number to be added. The operator
variable will have the text content of whatever operator button was
clicked.

Created event listener for the equals button, once clicked the display
value will be stored in num2, and clear the display.