Create functionality functions & edit operation listener
Completely changed the functionality.

Added a clearDisplay function to clear the display and checks to see if
the currentValue is not undefined. This was used to make sure that the
display would clear just once after an operation was performed so the
number buttons would append and not clear after the first click.

The function clearEverything is used once a user hits the clear button.
this will clear the display, the array, and make cuarrentValue
undefined.

Created a checkArray function that will check to see if any index of the
array includes an empty string.

Crated an event listener for the equals button. This will check to see
if the array has an empty string, and if so check each index to execute
the proper procedure.