Create delete function
This function deletes one number at a time if the user makes a mistake.
It will also check to see if the number is the result of an operation,
if so the delete button will not work, it will only work if the
currentValue variable is not equal to the display text.