Create default 0 & check for division by 0
Created a function to have a defalt zero show in the display, and also
checked to see when a number button is clicked, if the display equals 0,
and if the length is greater than or equal to 1. If so clear the display.

Create conditional statements to check if division by 0 is being
executed on the operator, and equals listeners. If so, display an error
message.