Edit delete button
When hitting the delete button, I needed to make sure that if there is
an error message that it cannot be deleted. Also, if the display is only
zero, you are unable to delete it. And if there is only one number left
to delete, once it is deleted the display will show 0;